+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/model/VirtualCluster.java
@@ -9,6 +9,7 @@ import java.io.UncheckedIOException;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
+import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
 import javax.net.ssl.SSLException;
@@ -24,9 +25,12 @@ import io.kroxylicious.proxy.config.tls.NettyKeyProvider;
 import io.kroxylicious.proxy.config.tls.NettyTrustProvider;
 import io.kroxylicious.proxy.config.tls.PlatformTrustProvider;
 import io.kroxylicious.proxy.config.tls.Tls;
+import io.kroxylicious.proxy.config.tls.TrustOptions;
 import io.kroxylicious.proxy.config.tls.TrustProvider;
 import io.kroxylicious.proxy.service.ClusterNetworkAddressConfigProvider;
 import io.kroxylicious.proxy.service.HostPort;
+
+import edu.umd.cs.findbugs.annotations.NonNull;
 
 @SuppressWarnings("OptionalUsedAsFieldOrParameterType")
 public class VirtualCluster implements ClusterNetworkAddressConfigProvider {
@@ -75,7 +79,12 @@ public class VirtualCluster implements ClusterNetworkAddressConfigProvider {
                                                  ClusterNetworkAddressConfigProvider clusterNetworkAddressConfigProvider,
                                                  Optional<Tls> tls) {
         try {
-            var downstreamTls = tls.map(tls1 -> " (TLS)").orElse("");
+            var downstreamTls = tls.map(t -> Optional.ofNullable(t.trust())
+                    .map(TrustProvider::trustOptions)
+                    .map(TrustOptions::toString)
+                    .orElse("-"))
+                    .map(options -> " (TLS: " + options + ")")
+                    .orElse("");
             HostPort downstreamBootstrap = clusterNetworkAddressConfigProvider.getClusterBootstrapAddress();
             var upstreamTls = targetCluster.tls().map(tls1 -> " (TLS)").orElse("");
             HostPort upstreamHostPort = targetCluster.bootstrapServersList().get(0);
@@ -180,7 +189,10 @@ public class VirtualCluster implements ClusterNetworkAddressConfigProvider {
     private Optional<SslContext> buildDownstreamSslContext() {
         return tls.map(tlsConfiguration -> {
             try {
-                return Optional.of(tlsConfiguration.key()).map(NettyKeyProvider::new).map(NettyKeyProvider::forServer).orElseThrow().build();
+                var sslContextBuilder = Optional.of(tlsConfiguration.key()).map(NettyKeyProvider::new).map(NettyKeyProvider::forServer)
+                        .orElseThrow();
+
+                return configureTrustProvider(tlsConfiguration).apply(sslContextBuilder).build();
             }
             catch (SSLException e) {
                 throw new UncheckedIOException(e);
@@ -194,8 +206,14 @@ public class VirtualCluster implements ClusterNetworkAddressConfigProvider {
                 var sslContextBuilder = Optional.ofNullable(targetClusterTls.key()).map(NettyKeyProvider::new).map(NettyKeyProvider::forClient)
                         .orElse(SslContextBuilder.forClient());
 
-                final TrustProvider trustProvider = Optional.ofNullable(targetClusterTls.trust()).orElse(PlatformTrustProvider.INSTANCE);
-                var withTrust = new NettyTrustProvider(trustProvider).apply(sslContextBuilder);
+                Optional.ofNullable(targetClusterTls.trust())
+                        .map(TrustProvider::trustOptions)
+                        .filter(Predicate.not(TrustOptions::forClient))
+                        .ifPresent(to -> {
+                            throw new IllegalStateException("Cannot apply trust options " + to + " to upstream (client) TLS.)");
+                        });
+
+                var withTrust = configureTrustProvider(targetClusterTls).apply(sslContextBuilder);
 
                 return withTrust.build();
             }
@@ -203,6 +221,12 @@ public class VirtualCluster implements ClusterNetworkAddressConfigProvider {
                 throw new UncheckedIOException(e);
             }
         });
+    }
+
+    @NonNull
+    private static NettyTrustProvider configureTrustProvider(Tls tlsConfiguration) {
+        final TrustProvider trustProvider = Optional.ofNullable(tlsConfiguration.trust()).orElse(PlatformTrustProvider.INSTANCE);
+        return new NettyTrustProvider(trustProvider);
     }
 
     private static void validatePortUsage(ClusterNetworkAddressConfigProvider clusterNetworkAddressConfigProvider) {